move ProxyDetector to public API
