+++ core/src/test/java/io/grpc/internal/ProxyDetectorImplTest.java
@@ -28,6 +28,7 @@ import static org.mockito.Mockito.when;
 
 import com.google.common.base.Supplier;
 import com.google.common.collect.ImmutableList;
+import io.grpc.ProxyDetector;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.PasswordAuthentication;
